requirements.txt
pandas==1.5.3
scipy==1.12.0
streamlit==1.32.1
plotly-express==0.4.1
